+++ src/Elasticsearch/State/CollectionProvider.php
@@ -24,6 +24,8 @@ use ApiPlatform\State\ProviderInterface;
 use Elastic\Elasticsearch\Client;
 use Elastic\Elasticsearch\Exception\ClientResponseException;
 use Elastic\Elasticsearch\Response\Elasticsearch;
+use Elasticsearch\Client as V7Client;
+use Elasticsearch\Common\Exceptions\Missing404Exception as V7Missing404Exception;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 
 /**
@@ -37,8 +39,13 @@ final class CollectionProvider implements ProviderInterface
     /**
      * @param RequestBodySearchCollectionExtensionInterface[] $collectionExtensions
      */
-    public function __construct(private readonly Client $client, private readonly ?DenormalizerInterface $denormalizer = null, private readonly ?Pagination $pagination = null, private readonly iterable $collectionExtensions = [], private readonly ?InflectorInterface $inflector = new Inflector())
-    {
+    public function __construct(
+        private readonly V7Client|Client $client, // @phpstan-ignore-line
+        private readonly ?DenormalizerInterface $denormalizer = null,
+        private readonly ?Pagination $pagination = null,
+        private readonly iterable $collectionExtensions = [],
+        private readonly ?InflectorInterface $inflector = new Inflector(),
+    ) {
     }
 
     /**
@@ -68,13 +75,15 @@ final class CollectionProvider implements ProviderInterface
         ];
 
         try {
-            $documents = $this->client->search($params);
+            $documents = $this->client->search($params); // @phpstan-ignore-line
+        } catch (V7Missing404Exception $e) { // @phpstan-ignore-line
+            throw new Error(status: $e->getCode(), detail: $e->getMessage(), title: $e->getMessage(), originalTrace: $e->getTrace()); // @phpstan-ignore-line
         } catch (ClientResponseException $e) {
             $response = $e->getResponse();
             throw new Error(status: $response->getStatusCode(), detail: (string) $response->getBody(), title: $response->getReasonPhrase(), originalTrace: $e->getTrace());
         }
 
-        if ($documents instanceof Elasticsearch) {
+        if (class_exists(Elasticsearch::class) && $documents instanceof Elasticsearch) {
             $documents = $documents->asArray();
         }