+++ src/Laravel/config/api-platform.php
@@ -11,6 +11,7 @@ return [
     'show_webby' => true,
 
     'routes' => [
+        'domain' => null,
         // Global middleware applied to every API Platform routes
         // 'middleware' => []
     ],
@@ -44,20 +45,20 @@ return [
         'pagination_enabled' => true,
         'pagination_partial' => false,
         'pagination_client_enabled' => false,
-		'pagination_client_items_per_page' => false,
-		'pagination_client_partial' => false,
-		'pagination_items_per_page' => 30,
-		'pagination_maximum_items_per_page' => 30,
+        'pagination_client_items_per_page' => false,
+        'pagination_client_partial' => false,
+        'pagination_items_per_page' => 30,
+        'pagination_maximum_items_per_page' => 30,
         'route_prefix' => '/api',
         'middleware' => [],
     ],
 
-	'pagination' => [
-		'page_parameter_name' => 'page',
-		'enabled_parameter_name' => 'pagination',
-		'items_per_page_parameter_name' => 'itemsPerPage',
-		'partial_parameter_name' => 'partial',
-	],
+    'pagination' => [
+        'page_parameter_name' => 'page',
+        'enabled_parameter_name' => 'pagination',
+        'items_per_page_parameter_name' => 'itemsPerPage',
+        'partial_parameter_name' => 'partial',
+    ],
 
     'graphql' => [
         'enabled' => false,
@@ -108,10 +109,17 @@ return [
         //]
     ],
 
+    // 'openapi' => [
+    //     'tags' => []
+    // ],
+
     'url_generation_strategy' => UrlGeneratorInterface::ABS_PATH,
 
     'serializer' => [
         'hydra_prefix' => false,
         // 'datetime_format' => \DateTimeInterface::RFC3339
-    ]
+    ],
+
+    // we recommend using "file" or "acpu"
+    'cache' => 'file'
 ];