+++ src/Metadata/ApiProperty.php
@@ -223,7 +223,7 @@ final class ApiProperty
         private array $extraProperties = [],
     ) {
         $this->types = \is_string($types) ? (array) $types : $types;
-        $this->serialize = \is_array($serialize) ? $serialize : (array) $serialize;
+        $this->serialize = \is_array($serialize) ? $serialize : [$serialize];
     }
 
     public function getProperty(): ?string