+++ src/Metadata/Util/ReflectionClassRecursiveIterator.php
@@ -22,12 +22,26 @@ namespace ApiPlatform\Metadata\Util;
  */
 final class ReflectionClassRecursiveIterator
 {
+    /**
+     * @var array<string, array<class-string, \ReflectionClass>>
+     */
+    private static array $localCache;
+
     private function __construct()
     {
     }
 
-    public static function getReflectionClassesFromDirectories(array $directories): \Iterator
+    /**
+     * @return array<class-string, \ReflectionClass>
+     */
+    public static function getReflectionClassesFromDirectories(array $directories): array
     {
+        $id = hash('xxh3', implode('', $directories));
+        if (isset(self::$localCache[$id])) {
+            return self::$localCache[$id];
+        }
+
+        $includedFiles = [];
         foreach ($directories as $path) {
             $iterator = new \RegexIterator(
                 new \RecursiveIteratorIterator(
@@ -61,12 +75,15 @@ final class ReflectionClassRecursiveIterator
         $sortedInterfaces = get_declared_interfaces();
         sort($sortedInterfaces);
         $declared = [...$sortedClasses, ...$sortedInterfaces];
+        $ret = [];
         foreach ($declared as $className) {
             $reflectionClass = new \ReflectionClass($className);
             $sourceFile = $reflectionClass->getFileName();
             if (isset($includedFiles[$sourceFile])) {
-                yield $className => $reflectionClass;
+                $ret[$className] = $reflectionClass;
             }
         }
+
+        return self::$localCache[$id] = $ret;
     }
 }