+++ src/OpenApi/Factory/OpenApiFactory.php
@@ -49,6 +49,7 @@ use ApiPlatform\OpenApi\Model\RequestBody;
 use ApiPlatform\OpenApi\Model\Response;
 use ApiPlatform\OpenApi\Model\SecurityScheme;
 use ApiPlatform\OpenApi\Model\Server;
+use ApiPlatform\OpenApi\Model\Tag;
 use ApiPlatform\OpenApi\OpenApi;
 use ApiPlatform\OpenApi\Options;
 use ApiPlatform\OpenApi\Serializer\NormalizeOperationNameTrait;
@@ -69,6 +70,7 @@ final class OpenApiFactory implements OpenApiFactoryInterface
     use TypeFactoryTrait;
 
     public const BASE_URL = 'base_url';
+    public const API_PLATFORM_TAG = 'x-apiplatform-tag';
     public const OVERRIDE_OPENAPI_RESPONSES = 'open_api_override_responses';
     private readonly Options $openApiOptions;
     private readonly PaginationOptions $paginationOptions;
@@ -93,6 +95,7 @@ final class OpenApiFactory implements OpenApiFactoryInterface
         ?Options $openApiOptions = null,
         ?PaginationOptions $paginationOptions = null,
         private readonly ?RouterInterface $router = null,
+        private readonly array $errorFormats = [],
     ) {
         $this->filterLocator = $filterLocator;
         $this->openApiOptions = $openApiOptions ?: new Options('API Platform');
@@ -101,6 +104,10 @@ final class OpenApiFactory implements OpenApiFactoryInterface
 
     /**
      * {@inheritdoc}
+     *
+     * You can filter openapi operations with the `x-apiplatform-tag` on an OpenApi Operation using the `filter_tags`.
+     *
+     * @param array{base_url?: string, filter_tags?: string[]}&array<string, mixed> $context
      */
     public function __invoke(array $context = []): OpenApi
     {
@@ -112,12 +119,12 @@ final class OpenApiFactory implements OpenApiFactoryInterface
         $paths = new Paths();
         $schemas = new \ArrayObject();
         $webhooks = new \ArrayObject();
+        $tags = [];
 
         foreach ($this->resourceNameCollectionFactory->create() as $resourceClass) {
             $resourceMetadataCollection = $this->resourceMetadataFactory->create($resourceClass);
-
             foreach ($resourceMetadataCollection as $resourceMetadata) {
-                $this->collectPaths($resourceMetadata, $resourceMetadataCollection, $paths, $schemas, $webhooks, $context);
+                $this->collectPaths($resourceMetadata, $resourceMetadataCollection, $paths, $schemas, $webhooks, $tags, $context);
             }
         }
 
@@ -127,6 +134,8 @@ final class OpenApiFactory implements OpenApiFactoryInterface
         foreach (array_keys($securitySchemes) as $key) {
             $securityRequirements[] = [$key => []];
         }
+
+        $globalTags = $this->openApiOptions->getTags() ?: array_values($tags) ?: [];
 
         return new OpenApi(
             $info,
@@ -142,17 +151,26 @@ final class OpenApiFactory implements OpenApiFactoryInterface
                 new \ArrayObject($securitySchemes)
             ),
             $securityRequirements,
-            [],
+            $globalTags,
             null,
             null,
             $webhooks
         );
     }
 
-    private function collectPaths(ApiResource $resource, ResourceMetadataCollection $resourceMetadataCollection, Paths $paths, \ArrayObject $schemas, \ArrayObject $webhooks, array $context = []): void
+    private function collectPaths(ApiResource $resource, ResourceMetadataCollection $resourceMetadataCollection, Paths $paths, \ArrayObject $schemas, \ArrayObject $webhooks, array &$tags, array $context = []): void
     {
         if (0 === $resource->getOperations()->count()) {
             return;
+        }
+
+        $defaultError = $this->getErrorResource($this->openApiOptions->getErrorResourceClass() ?? ApiResourceError::class);
+        $defaultValidationError = $this->getErrorResource($this->openApiOptions->getValidationErrorResourceClass() ?? ValidationException::class, 422, 'Unprocessable entity');
+
+        // This filters on our extension x-apiplatform-tag as the openapi operation tag is used for ordering operations
+        $filteredTags = $context['filter_tags'] ?? [];
+        if (!\is_array($filteredTags)) {
+            $filteredTags = [$filteredTags];
         }
 
         foreach ($resource->getOperations() as $operationName => $operation) {
@@ -166,6 +184,15 @@ final class OpenApiFactory implements OpenApiFactoryInterface
 
             // Operation ignored from OpenApi
             if ($operation instanceof HttpOperation && false === $openapiAttribute) {
+                continue;
+            }
+
+            $operationTag = ($openapiAttribute?->getExtensionProperties()[self::API_PLATFORM_TAG] ?? []);
+            if (!\is_array($operationTag)) {
+                $operationTag = [$operationTag];
+            }
+
+            if ($filteredTags && $filteredTags !== array_intersect($filteredTags, $operationTag)) {
                 continue;
             }
 
@@ -216,6 +243,10 @@ final class OpenApiFactory implements OpenApiFactoryInterface
                 servers: null !== $openapiOperation->getServers() ? $openapiOperation->getServers() : null,
                 extensionProperties: $openapiOperation->getExtensionProperties(),
             );
+
+            foreach ($openapiOperation->getTags() as $v) {
+                $tags[$v] = new Tag(name: $v, description: $resource->getDescription());
+            }
 
             [$requestMimeTypes, $responseMimeTypes] = $this->getMimeTypes($operation);
 
@@ -331,6 +362,7 @@ final class OpenApiFactory implements OpenApiFactoryInterface
             $openapiOperation = $openapiOperation->withParameters($openapiParameters);
             $existingResponses = $openapiOperation->getResponses() ?: [];
             $overrideResponses = $operation->getExtraProperties()[self::OVERRIDE_OPENAPI_RESPONSES] ?? $this->openApiOptions->getOverrideResponses();
+            $errors = null;
             if ($operation instanceof HttpOperation && null !== ($errors = $operation->getErrors())) {
                 /** @var array<class-string|string, Error> */
                 $errorOperations = [];
@@ -352,19 +384,24 @@ final class OpenApiFactory implements OpenApiFactoryInterface
                         $successStatus = (string) $operation->getStatus() ?: 201;
                         $openapiOperation = $this->buildOpenApiResponse($existingResponses, $successStatus, \sprintf('%s resource created', $resourceShortName), $openapiOperation, $operation, $responseMimeTypes, $operationOutputSchemas, $resourceMetadataCollection);
 
-                        $openapiOperation = $this->addOperationErrors($openapiOperation, [
-                            ApiResourceError::class => $this->getErrorResource(ApiResourceError::class)->withStatus(400)->withDescription('Invalid input'),
-                            ValidationException::class => $this->getErrorResource(ValidationException::class, 422, 'Unprocessable entity'), // we add defaults as ValidationException can not be installed
-                        ], $resourceMetadataCollection, $schema, $schemas, $operation);
+                        if (null === $errors) {
+                            $openapiOperation = $this->addOperationErrors($openapiOperation, [
+                                $defaultError->withStatus(400)->withDescription('Invalid input'),
+                                $defaultValidationError,
+                            ], $resourceMetadataCollection, $schema, $schemas, $operation);
+                        }
                         break;
                     case 'PATCH':
                     case 'PUT':
                         $successStatus = (string) $operation->getStatus() ?: 200;
                         $openapiOperation = $this->buildOpenApiResponse($existingResponses, $successStatus, \sprintf('%s resource updated', $resourceShortName), $openapiOperation, $operation, $responseMimeTypes, $operationOutputSchemas, $resourceMetadataCollection);
-                        $openapiOperation = $this->addOperationErrors($openapiOperation, [
-                            ApiResourceError::class => $this->getErrorResource(ApiResourceError::class)->withStatus(400)->withDescription('Invalid input'),
-                            ValidationException::class => $this->getErrorResource(ValidationException::class, 422, 'Unprocessable entity'), // we add defaults as ValidationException can not be installed
-                        ], $resourceMetadataCollection, $schema, $schemas, $operation);
+
+                        if (null === $errors) {
+                            $openapiOperation = $this->addOperationErrors($openapiOperation, [
+                                $defaultError->withStatus(400)->withDescription('Invalid input'),
+                                $defaultValidationError,
+                            ], $resourceMetadataCollection, $schema, $schemas, $operation);
+                        }
                         break;
                     case 'DELETE':
                         $successStatus = (string) $operation->getStatus() ?: 204;
@@ -375,13 +412,13 @@ final class OpenApiFactory implements OpenApiFactoryInterface
 
             if ($overrideResponses && !isset($existingResponses[403]) && $operation->getSecurity()) {
                 $openapiOperation = $this->addOperationErrors($openapiOperation, [
-                    ApiResourceError::class => $this->getErrorResource(ApiResourceError::class)->withStatus(403)->withDescription('Forbidden'),
+                    $defaultError->withStatus(403)->withDescription('Forbidden'),
                 ], $resourceMetadataCollection, $schema, $schemas, $operation);
             }
 
-            if ($overrideResponses && !$operation instanceof CollectionOperationInterface && 'POST' !== $operation->getMethod() && !isset($existingResponses[404])) {
+            if ($overrideResponses && !$operation instanceof CollectionOperationInterface && 'POST' !== $operation->getMethod() && !isset($existingResponses[404]) && null === $errors) {
                 $openapiOperation = $this->addOperationErrors($openapiOperation, [
-                    ApiResourceError::class => $this->getErrorResource(ApiResourceError::class)->withStatus(404)->withDescription('Not found'),
+                    $defaultError->withStatus(404)->withDescription('Not found'),
                 ], $resourceMetadataCollection, $schema, $schemas, $operation);
             }
 
@@ -848,8 +885,8 @@ final class OpenApiFactory implements OpenApiFactoryInterface
     }
 
     /**
-     * @param array<class-string, ErrorResource> $errors
-     * @param \ArrayObject<string, Schema>       $schemas
+     * @param ErrorResource[]              $errors
+     * @param \ArrayObject<string, Schema> $schemas
      */
     private function addOperationErrors(
         Operation $operation,
@@ -859,15 +896,20 @@ final class OpenApiFactory implements OpenApiFactoryInterface
         \ArrayObject $schemas,
         HttpOperation $originalOperation,
     ): Operation {
-        $defaultFormat = ['json' => ['application/problem+json']];
-        foreach ($errors as $error => $errorResource) {
-            $responseMimeTypes = $this->flattenMimeTypes($errorResource->getOutputFormats() ?: $defaultFormat);
+        foreach ($errors as $errorResource) {
+            $responseMimeTypes = $this->flattenMimeTypes($errorResource->getOutputFormats() ?: $this->errorFormats);
             foreach ($errorResource->getOperations() as $errorOperation) {
                 if (false === $errorOperation->getOpenApi()) {
                     continue;
                 }
 
-                $responseMimeTypes += $this->flattenMimeTypes($errorOperation->getOutputFormats() ?: $defaultFormat);
+                $responseMimeTypes += $this->flattenMimeTypes($errorOperation->getOutputFormats() ?: $this->errorFormats);
+            }
+
+            foreach ($responseMimeTypes as $mime => $format) {
+                if (!isset($this->errorFormats[$format])) {
+                    unset($responseMimeTypes[$mime]);
+                }
             }
 
             $operationErrorSchemas = [];
@@ -878,7 +920,7 @@ final class OpenApiFactory implements OpenApiFactoryInterface
             }
 
             if (!$status = $errorResource->getStatus()) {
-                throw new RuntimeException(\sprintf('The error class %s has no status defined, please either implement ProblemExceptionInterface, or make it an ErrorResource with a status', $error));
+                throw new RuntimeException(\sprintf('The error class "%s" has no status defined, please either implement ProblemExceptionInterface, or make it an ErrorResource with a status', $errorResource->getClass()));
             }
 
             $operation = $this->buildOpenApiResponse($operation->getResponses() ?: [], $status, $errorResource->getDescription() ?? '', $operation, $originalOperation, $responseMimeTypes, $operationErrorSchemas, $resourceMetadataCollection);
@@ -908,8 +950,10 @@ final class OpenApiFactory implements OpenApiFactoryInterface
             throw new RuntimeException(\sprintf('The error class "%s" does not implement "%s". Did you forget a use statement?', $error, ProblemExceptionInterface::class));
         }
 
+        $defaultErrorResourceClass = $this->openApiOptions->getErrorResourceClass() ?? ApiResourceError::class;
+
         try {
-            $errorResource = $this->resourceMetadataFactory->create($error)[0] ?? new ErrorResource(status: $status, description: $description, class: ApiResourceError::class);
+            $errorResource = $this->resourceMetadataFactory->create($error)[0] ?? new ErrorResource(status: $status, description: $description, class: $defaultErrorResourceClass);
             if (!($errorResource instanceof ErrorResource)) {
                 throw new RuntimeException(\sprintf('The error class %s is not an ErrorResource', $error));
             }
@@ -923,7 +967,7 @@ final class OpenApiFactory implements OpenApiFactoryInterface
                 $errorResource = $errorResource->withDescription($description);
             }
         } catch (ResourceClassNotFoundException|OperationNotFoundException) {
-            $errorResource = new ErrorResource(status: $status, description: $description, class: ApiResourceError::class);
+            $errorResource = new ErrorResource(status: $status, description: $description, class: $defaultErrorResourceClass);
         }
 
         if (!$errorResource->getClass()) {