+++ src/Symfony/EventListener/ErrorListener.php
@@ -94,7 +94,7 @@ final class ErrorListener extends SymfonyErrorListener
             $operation = $operation->withProvider('api_platform.state.error_provider');
         }
 
-        $normalizationContext = ($operation->getNormalizationContext() ?? []) + ($apiOperation?->getNormalizationContext() ?? []);
+        $normalizationContext = $operation->getNormalizationContext() ?? [];
         if (!($normalizationContext['api_error_resource'] ?? false)) {
             $normalizationContext += ['api_error_resource' => true];
         }